cleaned up complaints from jshint
JSHint complains and dies because the validation for options brings the
cyclomatic complexity of subaccounts.create to 6, where the settings say
the maximum should be 5. Refactored the validation of options into its
own function, but if there ever should be more required properties, then
we'll be right back here again.